DEVPROD-3971: create admin setting for hosts permanently exempt from sleep schedule
