+++ service/templates/admin.html
@@ -97,6 +97,7 @@ Admin Settings
 						<li class="link" ng-click="scrollTo('host_jasper')">Host Jasper</li>
 						<li class="link" ng-click="scrollTo('jira_notifications')">JIRA Custom Field Mappings</li>
 						<li class="link" ng-click="scrollTo('spawnhosts')">Spawn Hosts</li>
+						<li class="link" ng-click="scrollTo('sleep_schedule')">Sleep Schedules</li>
 						<li class="link" ng-click="scrollTo('tracer')">Tracer Config</li>
 						<li class="link" ng-click="scrollTo('project_creation')">Project Creation</li>
 						<li class="link" ng-click="scrollTo('github_check_run')">GitHub Check Run</li>
@@ -2427,9 +2428,28 @@ Admin Settings
 										<label>Unexpirable Spawn Hosts per user</label>
 										<input type="number" ng-model="Settings.spawnhost.unexpirable_hosts_per_user">
 									</md-input-container>
-									<md-input-container class="control" style="width:45%; margin-left:50px;">
+									<md-input-container class="control" style="width:45%;">
 										<label>Unexpirable Volumes per user</label>
 										<input type="number" ng-model="Settings.spawnhost.unexpirable_volumes_per_user">
+									</md-input-container>
+								</md-card-content>
+							</md-card>
+						</section>
+						<section layout="row" flex>
+							<md-card flex=50 id="sleep_schedule">
+								<md-card-title>
+									<md-card-title-text>
+										<span>Sleep Schedules</span>
+									</md-card-title-text>
+									<md-button ng-click="clearSection('sleep_schedule')">
+										<i class="fa fa-trash"></i>
+									</md-button>
+								</md-card-title>
+								<md-card-content>
+									<md-input-container class="control" style="width:45%;">
+										<label>Permanently Exempt Hosts</label>
+										<textarea ng-model="Settings.sleep_schedule.permanently_exempt_hosts" ng-list="&#10;" ng-trim="false" rows="3"
+											md-select-on-focus></textarea>
 									</md-input-container>
 								</md-card-content>
 							</md-card>
